requirements.txt
Django==5.2.7
djangorestframework==3.16.1
psycopg2-binary==2.9.9
whitenoise==6.7.0
gunicorn==23.0.0

